Create config-manager constant demo
A demonstration implementation of a potential config manager which would
be able to have all configuration options for robots (including button
mappings, constants, etc) has been created.

Not entirely sure on how to make sure that the json file is built onto
the RoboRIO nor how to access it once it is. Maybe @eandr127 could tell
me if this will work or not, which is why this is just a demo.