Made object-oriented base for fluid constants
In reference to #8

Was not entirely sure of how to implement the fluid constants system.
The best solution to me seems to be using object, which is a little
annoying given the memory issues we've had but this is a relatively
small number of objects each containing a small amount of data.

The other ways that I thought of were quite messy and would involve a
lot of work, which would go against the requirements of the issue this
is trying to solve.

If you could take a look at this @eandr127 before I continue and make
sure that this is an approach that works for you, that would be good.

If you wanted to access a constant, say the `OPERATOR_PRESS_A` constant,
you would do `Config.OPERATOR_PRESS_A.getValue()` in total.