Proposal for joystick_fluid
After reviewing @Zinka010's PR #28 for joystick_fluid, I had a couple of suggestions on how it might work.

I'm committing this to update the FluidConstants to have listeners for when their value changes and
other changes to the fluid constants in general in order to better allow for this to work.