Add windows CI for azure
Since the linux machine in CI uses a docker image to run the tests, it can behave
a little funky.

This is why I thought it would be a good idea to run our tests on both windows without any emulation
as well as using the emulated roboRIO platform.

We could add Mac OS but that seems pointless.