Working rumbler tests.
The rumbler tests are now functiontal though not ideal.

I mock the OI class which is something that I shouldn't do but I don't see much of a way around it since
if I don't mock it I end up with no way (or at least no way that I know of yet) to verify which instnace of a Joystick
methods are being called on.