Add test util, fix test architecture problems
I've now realized that having all the static stuff in robot and Config is actually really hurting
us since it's hard to ensure that all that stuff is reset properly when you want to make a new robot class instance.

I've made a hotfix for this stuff, but eventually it would be nice
to see if we can change the design of Config and Robot to avoid this sort of stuff.