Many changes, finalizing commands
Trying to finalize command structure for the intake and lift subsystems.
Most of the commands currently match Erik's document, some things I decided I'd
test before implementing to see if it would be beneficial or not (such as the button for slightly
descending the lift).

I made the tests pass (hopefully) and had a lot of cleanup and refactorings to do because of the document.