Improve with ETrigger
Needed something similar to `whenHeld()` for triggers, so I added `runWhileActive()` to the trigger by
createing a new class which inherits from Trigger caled ETrigger.

Problem with this right now is that it's the same logic as in `EButton` and I don't know how
I can make EButton which inherits from a Button which inherits from ETrigger instead of Trigger.