Raise arms during climb
Added a command in the climb command group to raise the arms.
Also took another look at the mechanism code surrounding arm movement and
improved it a little bit, since it didn't check the current arm state before
doing things and the commands didn't finish prematurely as they should
if the arms are already in the good state.