Resolve issue with always firing pneumatics
It is now possible to specify to always fire pneumatics, even if they are already in position,
however this isn't the default behaviour.
This should be fine because this has always been the case.

The problem with always firing them is that you need to turn them off after the time interval,
and that would mean either making the command always run for the move time even if the stuff
is already in position, or you would have to get creative.