Reversed piston solenoid directions
Turns out the whole time we had kForward set to move the arms up and kReverse to move them down,
which makes less sense so we've (Ryan and I) decided to reverse this. The forward solenoid id is now 3 instead of the 2 it was,
and the reverse is now 2 instead of 3.